+++ src/logic/service/turn.ts
@@ -6,7 +6,7 @@ import {
   SilverStar,
   TealStar,
 } from "@/ui/renderer/star";
-import { PositionId } from "../domain/position";
+import { Position, PositionId } from "../domain/position";
 import { generateStarPositions } from "../domain/setup";
 import { EventBus } from "../event/event-bus";
 import { BusEventData } from "../event/type";
@@ -14,6 +14,8 @@ import { gameStore } from "../store/game";
 import { starStore } from "../store/star";
 import { threeAppStore } from "../store/three-app";
 import { worldStore } from "../store/world";
+import { Meshes, meshesStore } from "../store/meshes";
+import { logger } from "@/shared-logic/logger";
 
 export class TurnService extends GameService {
   constructor(eventBus: EventBus) {
@@ -33,26 +35,48 @@ export class TurnService extends GameService {
     gameStore.nextTurn();
     worldStore.movePiece(startPos, endPos);
 
-    // add star to UI
-    const { stars } = worldStore.getState();
-    const starsPos = new Set(Object.keys(stars) as Array<PositionId>);
-
-    const { pieces: piecesMap } = worldStore.getState();
-    const piecesPos = new Set(Object.keys(piecesMap) as Array<PositionId>);
-    piecesPos.union(starsPos);
-
-    const starPosition = generateStarPositions(piecesPos, 1);
-    const StarKind = getRandomArrEl([
-      IndigoStar,
-      TealStar,
-      MagentaStar,
-      SilverStar,
-    ]);
-    const star = new StarKind(starPosition[0]);
+    const shouldGenerateStar = Math.random() < 0.75;
     const { threeApp } = threeAppStore.getState();
     if (!threeApp) throw new Error("");
-    threeApp.scene.getObjectByName("world")!.add(star);
-    worldStore.addStar(star, starPosition[0]);
+
+    if (shouldGenerateStar) {
+      // add star to UI
+      const { stars } = meshesStore.getState().meshes!;
+      const starsPos = new Set(Object.keys(stars) as Array<PositionId>);
+
+      const { pieces: piecesMap } = worldStore.getState();
+      const piecesPos = new Set(Object.keys(piecesMap) as Array<PositionId>);
+      const occupiedPos = piecesPos.union(starsPos);
+
+      const starPosition = generateStarPositions(occupiedPos, 1);
+      if (Position.equals(starPosition[0], endPos)) {
+        logger.log("OOPS", { starPosition, endPos });
+      }
+      const StarKind = getRandomArrEl([
+        IndigoStar,
+        TealStar,
+        MagentaStar,
+        SilverStar,
+      ]);
+      const star = new StarKind(starPosition[0]);
+      threeApp.scene.add(star);
+
+      const { meshes } = meshesStore.getState();
+      const nextMeshes: Meshes = {
+        ...meshes!,
+        stars: { ...meshes!.stars, [starPosition[0].toId()]: star },
+      };
+      meshesStore.setMeshes(nextMeshes);
+    }
+
+    logger.info("SCENE OBJECTS.........");
+    threeApp!.scene.traverse((obj) => {
+      if (obj.name.includes("star")) {
+        logger.warn(
+          `${obj.id}: ${obj.type} — ${obj.name}: ${obj.position.x},${obj.position.z}`,
+        );
+      }
+    });
   }
 
   private addStarToPlayer({