+++ src/ui/renderer/event-handler.ts
@@ -3,7 +3,7 @@ import { PerspectiveCamera, Vector2, WebGLRenderer } from "three";
 import { Piece } from "./piece";
 import { RaycasterHandler } from "./raycaster-handler";
 import { Tile } from "./tile";
-import { Meshes } from "@/logic/service/setup";
+import { Meshes } from "@/logic/store/meshes";
 
 export class EventHandler {
   private raycasterHandler: RaycasterHandler;