+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.1 - Fix bug
+- fix bug pop up on tile occupied by piece
+- cleaner production
+- small refactor
+
 # 0.1.0 - MVP 1
 - add music
 - add star